# release-24.3: sql: avoid unnecessary singleflight requests for role membership cache