… docs/generated/settings/settings-for-tenants.txt
… docs/generated/settings/settings.html
… pkg/bench/rttanalysis/testdata/benchmark_expectations
… pkg/sql/information_schema.go
… pkg/sql/logictest/testdata/logic_test/crdb_internal_default_privileges
… pkg/sql/logictest/testdata/logic_test/grant_table
… pkg/sql/logictest/testdata/logic_test/pg_catalog
… pkg/sql/pg_catalog.go
… pkg/sql/user_test.go